requirements.txt
streamlit==1.40.0
pandas==2.2.3
pymilvus==2.4.9
langchain==0.3.7
openai==1.54.3
python-dotenv==1.0.1